requirements.txt
flask
flask_sqlalchemy
pyjwt
cryptography
